Added subprocesses checking
This checks if you want to have an RP when the desired [Program] is open.
If left [Optional] or None, it will be a default RP that runs with your
configurations and automatically show up.

Addendum:
subprocesses is in the while loop to keep checking if there are changes
in the running processes. If the [Program] is detected, then RP runs,
if [Program] is closed or is not open, RP closes.